rename scattering/calculations.py and inference/functional_api.py both to interface.py
