+++ holopy/scattering/interface.py
