+++ holopy/scattering/tests/test_basics.py
@@ -25,7 +25,7 @@ from holopy.core.tests.common import assert_obj_close
 from holopy.core.metadata import update_metadata
 from holopy.scattering.scatterer import Sphere, Difference
 from holopy.scattering.theory import Mie
-from holopy.scattering.calculations import calc_intensity, calc_holo, calc_field
+from holopy.scattering.interface import calc_intensity, calc_holo, calc_field
 
 # small tests against results from the previous version of holopy
